Thanish's 4th Random forest
The code is present in RF_sub_4.R and the respective predicted csv file is in RF_predicted_facies_4.csv. In this approach I have used "Shankle" as my validation set after learning from the previous mistake not to use random split for validation and also F1 as the validation metric with the micro average.